requirements.txt
streamlit
pdfplumber
pandas
pydantic
tenacity
google-genai
fpdf2